requirements.txt
alembic==1.14.1
amqp==5.3.1
aniso8601==10.0.0
asgiref==3.8.1
bcrypt==4.2.1
billiard==4.2.1
black==25.1.0
blinker==1.9.0
cairocffi==1.7.1
CairoSVG==2.7.1
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.7.0
cssselect2==0.8.0
defusedxml==0.7.1
Deprecated==1.2.18
Django==5.1.3
django-extensions==3.2.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.7.1
Flask-Limiter==3.10.1
Flask-Mail==0.10.0
flask-marshmallow==1.3.0
Flask-Migrate==4.1.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
html5lib==1.1
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
kombu==5.4.2
limits==4.0.1
Mako==1.3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
marshmallow-enum==1.5.1
marshmallow-sqlalchemy==1.4.1
mdurl==0.1.2
mypy-extensions==1.0.0
ordered-set==4.1.0
packaging==24.2
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.6
pluggy==1.5.0
prompt_toolkit==3.0.50
psycopg==3.2.3
psycopg2-binary==2.9.10
pycparser==2.22
Pygments==2.19.1
PyJWT==2.10.0
pyphen==0.17.2
pytest==8.3.5
pytest-cov==6.0.0
pytest-flask==1.3.0
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
redis==5.2.1
requests==2.32.3
rich==13.9.4
setuptools==76.1.0
six==1.17.0
SQLAlchemy==2.0.38
sqlparse==0.5.2
tinycss2==1.4.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
WeasyPrint==52.5
webencodings==0.5.1
Werkzeug==3.1.3
wkhtmltopdf==0.2
wrapt==1.17.2
gunicorn==23.0.0